chore: add more schema generation tests
