+++ src/test/java/com/github/victools/jsonschema/generator/SchemaGeneratorSimpleTypesTest.java
@@ -1,0 +1,86 @@
+/*
+ * Copyright 2019 VicTools.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.github.victools.jsonschema.generator;
+
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import junitparams.JUnitParamsRunner;
+import junitparams.Parameters;
+import org.junit.Assert;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+/**
+ * Test for {@link SchemaGenerator} class.
+ */
+@RunWith(JUnitParamsRunner.class)
+public class SchemaGeneratorSimpleTypesTest {
+
+    Object parametersForTestGenerateSchema_SimpleType() {
+        return new Object[][]{
+            {Object.class, null},
+            {String.class, SchemaConstants.TAG_TYPE_STRING},
+            {Character.class, SchemaConstants.TAG_TYPE_STRING},
+            {char.class, SchemaConstants.TAG_TYPE_STRING},
+            {CharSequence.class, SchemaConstants.TAG_TYPE_STRING},
+            {Boolean.class, SchemaConstants.TAG_TYPE_BOOLEAN},
+            {boolean.class, SchemaConstants.TAG_TYPE_BOOLEAN},
+            {Integer.class, SchemaConstants.TAG_TYPE_INTEGER},
+            {int.class, SchemaConstants.TAG_TYPE_INTEGER},
+            {Long.class, SchemaConstants.TAG_TYPE_INTEGER},
+            {long.class, SchemaConstants.TAG_TYPE_INTEGER},
+            {Short.class, SchemaConstants.TAG_TYPE_INTEGER},
+            {short.class, SchemaConstants.TAG_TYPE_INTEGER},
+            {Byte.class, SchemaConstants.TAG_TYPE_INTEGER},
+            {byte.class, SchemaConstants.TAG_TYPE_INTEGER},
+            {Double.class, SchemaConstants.TAG_TYPE_NUMBER},
+            {double.class, SchemaConstants.TAG_TYPE_NUMBER},
+            {Float.class, SchemaConstants.TAG_TYPE_NUMBER},
+            {float.class, SchemaConstants.TAG_TYPE_NUMBER}
+        };
+    }
+
+    @Test
+    @Parameters
+    public void testGenerateSchema_SimpleType(Class<?> targetType, String expectedJsonSchemaType) throws Exception {
+        SchemaGeneratorConfig config = new SchemaGeneratorConfigBuilder(new ObjectMapper()).build();
+        SchemaGenerator generator = new SchemaGenerator(config);
+        JsonNode result = generator.generateSchema(targetType);
+        if (expectedJsonSchemaType == null) {
+            Assert.assertTrue(result.isEmpty());
+        } else {
+            Assert.assertEquals(1, result.size());
+            Assert.assertEquals(expectedJsonSchemaType, result.get(SchemaConstants.TAG_TYPE).asText());
+        }
+    }
+
+    @Test
+    @Parameters(method = "parametersForTestGenerateSchema_SimpleType")
+    public void testGenerateSchema_SimpleType_withAdditionalPropertiesOption(Class<?> targetType, String expectedJsonSchemaType) throws Exception {
+        SchemaGeneratorConfig config = new SchemaGeneratorConfigBuilder(new ObjectMapper())
+                .with(Option.FORBIDDEN_ADDITIONAL_PROPERTIES_BY_DEFAULT)
+                .build();
+        SchemaGenerator generator = new SchemaGenerator(config);
+        JsonNode result = generator.generateSchema(targetType);
+        if (expectedJsonSchemaType == null) {
+            Assert.assertTrue(result.isEmpty());
+        } else {
+            Assert.assertEquals(1, result.size());
+            Assert.assertEquals(expectedJsonSchemaType, result.get(SchemaConstants.TAG_TYPE).asText());
+        }
+    }
+}
